requirements.txt
cryptography==41.0.1
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
psycopg2==2.9.6
pycryptodome==3.18.0
PyJWT==2.7.0
requests==2.31.0

